requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
attrs==18.2.0
Automat==0.7.0
black==18.9b0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
cssselect==1.0.3
Django==2.1.3
djangorestframework==3.9.0
drf-yasg==1.11.0
flex==6.13.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
lazy-object-proxy==1.3.1
logzero==1.5.0
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
parsel==1.5.1
pip-autoremove==0.9.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.1.1
pyOpenSSL==18.0.0
pytz==2018.7
PyYAML==3.13
queuelib==1.5.0
requests==2.20.1
rfc3987==1.3.8
ruamel.yaml==0.15.79
schedule==0.5.0
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
toml==0.10.0
Twisted==18.9.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
w3lib==1.19.0
wrapt==1.10.11
zope.interface==4.6.0
